refactored out util classes
