+++ src/easyscience/fitting/minimizers/bumps_utils/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2026 EasyScience contributors <https://github.com/easyscience>
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .eval_counter import EvalCounter
+from .progress_monitor import BumpsProgressMonitor
+
+__all__ = ['BumpsProgressMonitor', 'EvalCounter']
